pyproject.toml
[project]
name = "arena"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "loguru>=0.7.3",
    "numpy>=2.3.5",
    "pillow>=12.0.0",
    "simple-parsing>=0.1.7",
    "torch>=2.9.1",
]

[project.scripts]
arena-run = "arena.run:main"
arena-render = "arena.render:main"
